Add an auto-login module and make it the default.
This module simply forces a login of the defult user when
constructed, thus removing the current manual login requirement.
Also have the IDE do a load info call immediately so that it
doesn't show the login prompt.

*******************    DEBUG  USE  ONLY    **********************

This will make debugging the IDE much faster since you don't need
to login every time to want to refresh the page.