Have make dev build the applet
The applet is needed for development (else the IDE falls over at load) so this
makes the initial setup easier. This patch also make the applet depend on the
keystore, further simplifying setup.

Change the README to reflect the above changes -- basically stop telling the
user that they need to worry about building the applet.