Add some actual user settings.
Somewhat minimal at the moment, but the basics are there. This also
gives an idea of how they're going to be implemented.

Things that need to be added include support for default values,
actually saving the settings somewhere, better interactions with
the main user object and possibly setting dependancy (only showing
a setting once another has a certain value).