Allow settings to have a (blocking) "Please select" option.
This adds an extra option to the dropdown that gets removed as soon as
the user changes the setting's value. The settings cannot be saved until
all such settings have been changed to meaningful values.

This therefore can be used to force the user to make a chioce about a
setting. Once the user has saved a preference for a setting this option
will not show itself.