Automatically offer Java after the fourth checkout
If the user has not chosen about the use of java (but their browser claims
to support it) after their fourth checkout then force them to make a choice
whether or not to use it via the settings page. Make the java use setting
use the new "Please select" option to force the choice.