Implement RSS feed parsing using SimplePie.
Add notes to the README on how to get SimplePie installed.
Fill out the implementations of getRSS and GetRecentPosts in Feeds.
Add the include of SimplePie to the top of Feeds.

Add a couple of config options to determine the locaiton (defaults
to /tmp/ide-feed-cache) of the cache that SimplePie uses, and how
long it keeps it for.