Revert "Revert "Move the main html file out of the way.""
This reverts commit 0d927c61533be52105283f6ea1c303a7c5c3e569.

There's a many reasons I moved it out of the way, and the reason for the
revert is wrong...

I planned that we'd use .htaccess files to block
indexes on the web/ folder, and also probably block direct access to
both the .html files in there. Also it seemed to make more sense to
have it there, since that's the `web` folder.. and where the
login.html file is.

The other reason that I moved it was so that it was more immediately
clear to the webserver or new developer which the index file is (yeah,
I know that it can be configured).

We don't want users to be able to get to the index.html file at all
since that would utterly circumvent the separate login situation that
I've now merged in. While this isn't an issue in development (in fact
I agree that it might be simpler to just fire up the index.html file
in this case) it could get in the way in production, maybe.