Add a config option for user admin.
The default user's admin status can now be controlled by
a config switch interpreted as a boolean. I've added a
placeholder function in the LDAP stuff, which, hopefully,
someone who understands that side of things can flesh out.

I've intentionally not cached this in the auth token so that
a user who is stripped of their admin rights cannot abuse an
old token -- being an IDE admin give significant rights to all
sorts of things.