Make getCurrentRevision work on bare repos
This is needed by the upcoming poll module, since it will use the
master repos as its source of information in order to ensure that
it can quickly and easily get the most recent information.

Ideally, wherever possible, commands should be set to use the repo
rather than being run in the working copy. This ensures that we
don't accidentally pick up information from the host IDE repo, as
was happening in this case.