Fix loading of autosaved file data.
The autosave data is now sent up when the call is for the 'HEAD'
revision, otherwise it's set to null. This allows us to figure out
if the file is modified and notify the user appropriately, notably
on file close. Currently the file tab doesn't show this on autosave
load, but I plan to fix that.