Fixed a bug where the automagic config file would contain syntax errors.
This was as a result of using the built-in 'echo' in the Makefile. In
some shells, the built-in echo is different to /bin/echo in that it does
not accept the -n option, which is an option to cause echo not to emit a
newline. This was causing extraneous -ns and newlines in the Makefile on
some system configurations, in particular on Darwin.

This changes the Makefile to invoke /bin/echo explicitly rather than
using the built-in echo, which is guaranteed by POSIX-conforming systems
to accept the -n option.

Signed-off-by: Alastair Lynn <arplynn@gmail.com>