Added support for attaching a pyenv zip inside an existing zip.
This uses zip -r to attach the pyenv zip, specified in a config option,
into the archive after it's been generated by git archive.

In the next commit, it will be specified to use -0 to force non-compression:
since the pyenv zip is already compressed, further compression will not
be much of a win and will only slow down decompression.

Signed-off-by: Alastair Lynn <arplynn@gmail.com>