Hotfix for a bug with deauthentication when client and server are desynced.
The issue was that occasionally cookies would be kept longer in the client
than the server expected. This is highly problematic because if the client
clicks the 'log out' button, the server doesn't know what is going on and
bails.

The server now checks, when receiving the message to deauthenticate, that the
user is logged in in the first place - if not, it fails silently. The end
result is the same in either case - both the client and the server believe
deauthentication has occurred.

Signed-off-by: Alastair Lynn <arplynn@gmail.com>