Fixed a bug which prevented IE8 checkout out robot.zip in certain cases.
In some cases (such as the machines at Taunton's, apparently), IE8 was giving
badly-worded error messages on checkout. This was apparently due to a
Microsoft 'feature' which would cause IE8 to die if told not to cache files
while connected over SSL.

This is fixed with an htaccess file which tells the client to cache zips for
six seconds, and to send a content-disposition to ensure it's opened as a
download.

Signed-off-by: Alastair Lynn <arplynn@gmail.com>