Refactor to ensure shell-safe variables.
We've agreed that shell-safe variables will start '$s_', and there's
now a test that insists this be the case.
This commit renames all the existing safe ones, as well as ensuring
that all the other ones are safe too.

This is in reponse to the number of times we've had issues with
spacey paths