Hotfix: prevent items not in the index being wiped on update.
This enables the creation of folders within a working copy.

Previously when the repo updated the only changes it preserved were
those to indexed files. This is undesired behaviour, as it is possible
that a file has been autosaved, or a folder created, that is not yet
in the index (folders aren't indexed at all of course).

This patch changes the behaviour of GitRepository->unstagedChanges
to include all files present, indexed or otherwise. The previous
behaviour is available via an argument that defaults to false.
This patch also adds GitRepository->listFolders, which is used to
preserve empty folders in the working copy.

These new functions are now used in ProjectManager->updateRepository
so that these un-indexed files are not lost.