Hotfix: #701: Stop autosave moving code around.
Previously the autosave function called to get the current code form
the editor window, even if it was not the tab in focus. Now it checks
whether it's the tab in focus, and only updates from the EditArea if
this is true. Otherwise it uses the code it's got stashed away as the
code to autosave.