Hotfix: Fix wierd bug when exporting a spacey project.
I'm still not sure what the bug actually was, but I'm pretty sure
that sam's earlier spacey path fixing patches weren't sufficient.
I've remove the shell_exec("rm -rf $projdir"); line, which I think
was to blame, and switched to using a clean tmpdir for each checkout.
This removes the need for most of the checking/removing of files &
folders, and simplifies things.

I'm also switching to using the commit hash in the path that the
served robot.zip has. This, in combination with the use of a fresh
tmpdir each time, removes all posibilities of race conditions since
any call for an export has unique paths at (almost) all times. The
only expection to this is when there are two calls for the same
revision to be exported, but since this represents the same code being
exported the resulting overwrite of the export zip shouldn't casue
any issues. There's scope to detect the presence of the existence of
that cached zip from its hash (and thus avoid re-zipping), but that's
not for a hotfix.