Fixed a bug where getCurrentRevision() would return extraneous characters.
Once upon a happy time, gitExecute() would trim all whitespace from the start
and end of the output. However, this caused some problems when the whitespace
was necessary - for instance, with projects the names of which ended or began
with whitespace. gitExecute was thusly modified not to strip newlines.

However, this was a problem because `git describe' DOES return an extra
newline after the commit hash or tag reference.

Since these, we can be reasonably sure, will not start or end with newlines,
we now trim the return value from gitExecute.

Signed-off-by: Alastair Lynn <arplynn@gmail.com>