Fix an Out of Memory error in IE7
We weren't binding to all of the method parameters, so the last one
was picking up the event object being passed when you clicked the
button. This fix binds false to the last parameter, thus avoiding
this.

This needs review for next year, since we shouldn't assume a value
for the autosave paramter. See http://trac.srobo.org/ticket/735.