code-review: Improved the docstring for 'saveTeamName' in 'AdminModule'.
The previous docstring was somewhat non-representative of actual behaviour.

For the record, Sam and I have precisely dick all idea what the hell the feeds
stuff actually does. We will be assembling a crack team to dig down, but are
afraid of delving too deep and finding that feeds.php is full of balrogs.

This would not be entirely conducive to the continued furtherment of SR.

This is part of a code review conducted on 2011-04-05.

Signed-off-by: Sam Phippen <samphippen@googlemail.com>
Signed-off-by: Alistair Lynn <arplynn@gmail.com>