code-review: Improved functionality of 'putFeeds'.
Previously, this returned false in case of errors and the length
of the string it encoded in case of success. This was madness.

This has been changed to throw an exception in case of errors and
not to return any value (that is, to return null) in all other cases.

In the process, the code has been made more readable.

This is part of a code review conducted on 2011-04-05.

Signed-off-by: Sam Phippen <samphippen@googlemail.com>
Signed-off-by: Alistair Lynn <arplynn@gmail.com>