Revert "Undid moving of tests into tests/php."
This reverts commit bb2c83e7c29bdd01efc50c0d5b36acf68f9a71ff.
Master never had the tests moved, so this commit actually turned
out to be a bit destructive. I can't be bothered to re-run the
rebase to remove it, a revert will do.

This brings back the changes made to the project creation test
on the master branch. Notably testing for a slew of actually
complicated names.