Move the PyLint specific linting into its own lint class.
Thus making file->lintFile much simpler. Note that this method is
still responsible for creation and cleanup of the temporary directory
the linting is done in, and for moving the reference file into place.
The linting class *only* does the linting.

I'm intending to add more linters that look for other things, and
there's no point in each of them doing temp folder setup etc. This
also enables them to be very simple.