Implement import linting and recursive linting.
Import checking is achieved via a wrapper around python's modulefinder
module, and a custom AST visitor to get file positions. These output
a json dictionary of {file: {bad_import: line}} which is converted to
LintMessages by the php wrapper. Imports are checked for locality
by comparing their paths to the dirname from the primary file.

The list of files produced as the keys to the import linting dictionary
is then used to have those files parsed also, enabling the apparently
recursive linting.

Note that this messes with the check-all button on the errors page,
but we can live with that. A more pressing concern is that currently
only the autosave (or otherwise) state of the primary file is examined
and dealt with.