Make double-clicking on an error take you to it in the file.
You can now double click an item on the error page and it will take
you to the right line in the file the error is in, opening it first
if this is needed.

Make the edit_file method on the edit page return the etab for the
file in question and add a setSelectionRange method to the etab.
These are needed since the errors page doesn't know (or want to)
about the internals of either the errors page or editArea.

Note that this version of setSelectionRange uses line numbers, not
string positions for the same reasons. It converts these for use.