Work around pylint's failings.
If we call pylint on multiple files and one of them causes it to bomb
we lose all the information. This commit separates the files so they're
called one at a time, and then collates the results at the end.

If they all fall over this is reported also, but if any one succeeds
then the other failures are ignored.