Better reporting for a known fail case in pylint.
If there's an IndentationError thrown inside an import then pylint
is known to bomb. This commit allows us to detect such cases and
report what information we can back to the user so they can fix it.

Prior to this they'd simply get a fail message that didn't show that
it was their code at fault.