Fix #767 by refusing to validate usernames containing whitespace.
Requests anywhere else for usernames containing whitespace are unlikely
now that this is fixed, and since they'll likely work anyway (as this
was previously) I don't think this matters much. There might be cause
for this to be in the parent class, but I think this will do for now.