Fix #711 by creating the export applet on-demand too.
Previously the applet was only available if it had been created during
init. Now this is delegated to a helper, which is called during init,
that can be called repeatedly and simply creates the applet if it's
not already present but should be.

This allows for users enabling the java applet to have it created the
next time they try to export their project.