Fix #566: Users can now open old revisions from the log page.
Add a new 'View' button to the log page, which calls for the file
to be opened at the selected revision.

Modify the edit page so that if a call is made for an already open
file to be opened at a specific revision that the file tries to
change to being at that revision, depending on its modification
status.