Fix the handling of the number of pages (#577, part 2).
The server now reports the number of pages as expected, again using
the size of the current request as a gauge. This is now correctly
presented to the user at the top of the page.

Also rename the variable in JS from overflow to pageCount, which is
much, much, clearer.