Tidy the project options box.
Comment out the two disabled buttons that just look weird. They've
been disabled for the last couple of years, and it looks a bit
rubbish that we've not yet done something about them.

Improve the CSS for the buttons and text in the box also, such that
the buttons have a fixed size (much cleaner than the hodge-podge
there used to be) and line the word "Project" up with the LHS of
the buttons (in Firefox at least).