Don't reload HEAD when opening a file from the errors page.
Previously we always reloaded the file. Now we avoid the wrapped
revision loading mechanism so that we can check we're not reloading
the file pointlessly. If the user has modified the file, it's most
likely that the file's state is the one that was checked. This is
something that should be fixed at some point, but will do for now.