Fix handling of deleted files when updating a user repo.
This is a partial fix for #986, but the commit step also needs work.
We use the fact that the reading of a deleted file fails (returning
false) to notice its deletion and restore this after the update.
This is less than ideal, as the user might end up deleting a file that
has changed since they hit delete, however the chances of this are
small and since everything is versioned it probably doesn't matter
too much. We should consider detecting such issues and warning the
user at some point.