Acquire a lock on a git repository before using it
Requests to modify the git repository get interleaved.  Therefore,
there needs to be locking on the git repos.  This patch creates a lock
in the GitRepository constructor, and frees it in the destructor.

Right now, this results in deadlock because two things attempt to open
the git repository at the same time...