Ensure that failed commits are reported.
This should fix #987, and makes the IDE a lot more usable. For the
moment I'm returning false in the endpoing when the commit fails,
which I wanted to avoid, but for the moment this is the fastest way
to ship a fix. This relies on 03ec5891, which isn't yet merged into
this branch.

We need to do a proper review of the error handling routes at some
point.