Fix the deadlocking issue by maintaining a list of handles.
Add a factory method that caches the repos handles as they're created,
thus ensuring that we won't try to make more than one for a given repo
on each request. This isn't especially efficient, since each handle
lasts the entire length of the request, which is probably far more than
we actually need.