Rewrite the logging as a class, add log levels.
Having this in a static helper class seemed sane, and makes it a bit
more testable. Add a test for the level checking.

Using the PHP in-built log levels allows the ini-parsing of the config
files to automagically convert into the equivalent defined value.
However, we do then need a conversion from those values to printable
strings.

Add suitable logging levels to all the existing logging calls, and
set a default value of LOG_DEBUG, to match the current debug setting.