Add the start of backend support for Team Status.
This adds support for reading & writing to the status files, and
provides an abstraction that allows the underlying format to be changed
if needed. Support for file uploads is possible via an endpoint, but
needs its own php script to handle it, since the input format will
need to be different.