Add support for file uploads on the Team Status page.
The backend to this intentionally mimics the existing control.php
interface, but falls back to using $_POST variables for everything
since I couldn't find a way to send other data into the iframe.

The frontend uses a hidden iframe to target the form at, and listens
for the its event to gather the data which is returned. This mechanism
is so far only tested in Firefox.