Resize the uploaded images and convert to PNG.
For simplicity we want all the images to be the same size and type
so resize and convert them on upload. The backing class for this I've
pretty much lifted from another of my projects, and re-worked a bit
to use PHP 5 OO and Exceptions.

Note that this commit changes the config option for the status images
folde from being team.status_image_dir to team.status_images.dir so that
we can have .width and .height as well.