Add support for Python tests, and a basic python test.
Add a new python test runner (which was actually referenced, but
not preset, in my previous commit, 4a6614d2b5) that runs PyUnit
tests via its command line interface. This unfortunately requires
the presence of __init__.py files in the tree, prizes for anyone that
fixes this.

This also adds a utils class for running python tests against the
IDE, which includes a bunch of assertions that are nicer than those
in raw PyUnit as well as various IDE related helpers. There's also
a skip method in there too.

The config file included in this commit enables an arbitrary instance
of the IDE to be tested, defaulting to the live one. This can be easily
overridden by creating a localconfig beside it, which it will try to
import overrides from.

The test that this adds simply checks that it gets back a display-name
and token from the auth endpoint, and provides basic testing for much
of the framework I'm adding here.