Add support for storing review status.
This is very basic at the moment, and I'd like to store much more
information than this at some point, but for the moment this will do.
This will be used on the Admin page, when I get around to committing
it, and to display feedback information on the Team Status page.

Also a collection of tests, and update the existing ones to match the
slight change in the underlying format.