Replace the Admin page with one that validates Team Status.
The old page didn't do anything now that we've moved to Team Status,
so replace it with a new one. The validation mechanism provided here
is pretty simple, and is deliberately done per-team. The backend
is pretty much a lightweight wrapper onto the existing TeamStatus
class which does most of the actual logic handling.