Implement feedback on the Team Status page.
This consists of getting the backend to report the information from
the TeamStatus wrapper and then having the front-end display this.
We clear the review-status after they make a save, and then bring it
back from a fresh load of the information. We also now clear the inputs
when the page loads, so they won't have a flash of data from an old
page lying around.