Fix a memory leak in the Team Status page.
We weren't actually disconnecting any of the signals in the array
that we had so dutifully collected. I have a feeling that this error
may be present on a number of other pages...
Note that we now need to have the connection to the save button restored
when the page gets inited, otherwise the button gets disconnected the
first time it's closed (correctly) and then never connected again.