Rewrite the checkout building code as a helper class.
This moves the code to build the zip into a helper class so that the
project module can just get on with serving it.

Improve the separation between pyenv and the IDE further by requiring
the zip building script to be callable directly, and move its location
into a config option. The location of this script is also built by
a helper so that if needed it can change on the fly.

This is still very much based on the pre-existing code from the proj
module, but has been tweaked to work in a helper, and the redundant
helpers (fastwrap and completeArchive) are removed. If there's ever
a need to use fastwrap again, it should be easy enough to make it use
the new config setup.