Ensure that file-locks can't deadlock threads.
Have our attempts to lock files do so in a non-blocking manner, but
retry for a configurable amount of time. This ensures that two
simultaneous requests cannot possibly deadlock on each other, but
it does mean that they may instead fail (which is no worse a position,
since the user is now informed of this, and can retry).