Connect up the Team Status save to Pipebot.
Grab the latest pipebot so that the say script works and change the
condition in announce since the config reader doesn't do bools.

Have the team module output a useful message when the team saves
their status or image. We might need to add some actual change
detection in here, but we'll see how it goes.