Make it possible to serve each team a different pyenv version.
Add a config option that specifies which version of libRobot (currently
pyenv) a team should have included in their zips, with a fallback to
the current version if none is specified.

TODO: Tests, and handling of the case where a pyenv revision that
doesn't exist in the repo has been specified.