Avoid calling file/log twice on openning a file.
This bug has been in the IDE for as long as I can remember, and I've
been meaning to try to fix it for ages. The issue here is that
_update_contents (which calls for file/log) is called in a number of
places, many of which can do so simultaneously. This commit only partly
mitigates one of these cases, hence 'avoid' rather than 'fix'.

The particular case that this avoids is that when openning a file,
we both requst the contents of the file, which will eventually call
_recv_contents which calls to set the content of the file in EditArea,
and at the same time focus the newly created tab. The focussing of
course calls to set the contents of EA so that they're correct, having
previously come from an unknown state.

The other case I'm aware of where file/log is called twice is when
switching to an already open file. This is caused by the open call
also setting the desired revision of the file, which causes the file's
contents to be re-loaded from the backend.