Fix #1147: Ensure that the admin page reviews the right team.
This was a bug where the admin page used the IDE global team value
as the review target, thus causing potentially spurious reviews to
happen. Thankfully the backend has some protection against this, but
it also made it impossible to review a team's status if the reviewer
wasn't a member of the team.

Implemented fix by storing the current review team as a member, and
clearing this whenever it might change or become invalid.