Add a test for issue #1058.
This is the first 'real' test using the python/http interface to
the IDE, and, while I believe that the framework is reliable, is
unfortunately remarkably dependant on the conditions that make this
issue show itself. In particular, those are that the file-lock timeout
is short, and that the backend is given three requests in very quick
succession, which then clash.

That said, with the default lock timeout (200ms) in place, this test
repeatably fails while using my ide clone on optimus as the test subject.
Testing against localhost, however, seems less able to prove failures.