Optimise file linting.
Delay the initial repsitory interaction until after the pre-checks
and avoid other actions on the user's repo. Since we may be editing
it (possibly checking out files), this might as well occur on the
copied repo, rather than the original so as to minimise our locking
requirements. It also means that we don't need to worry about undo-ing
the checkout.