Fix #996: Add a cron job that tidies the export zips.
This uses the cron mechanism I just built, and has the proj module
subscribe a cron handler that deletes any zips in the zipurl folder
that haven't been accessed in the last 12 hours. This isn't a complete
fix, as there's still nothing that's calling the cron endpoint, but
it makes it trivial to do the rest.