Don't offer the user Java on the fourth checkout.
This confuses the user, and we'd like to remove the applet eventually
anyway. By not tell the user about it, they're less likely to miss
it when it is removed, and in the mean-time users that really want
to use it can continue to do so by setting the setting as they
always have.