Fix #1365: Project creation needs suitable author details.
This issue became particularly evident on Fedora 17 where the
version of git included now fails at this point rather than using
a value constructed from the system's own name. The change introduced
here therefore supplies these values directly, and makes them
configurable.

As a result of these values not being constructed from the system's
name, the current filtering of the log to exclude commits by the
log endpoint will break. I'm not certain whether or not there's a
need to keep doing this filtering, but either way, it'll need changing
to cope.