Also add user definition environment to stash.
For reasons that I can sort-of see, but don't care about, git needs
user details to stash. Unfortunately, stash is used as a bit of an
internal hack anyway, so just use the system details. We probably
could fix this by associating the user details with the repo in the
constructor, but this fixes it for now.