Remove the hiding of the initial commit from the log.
This was a workaround to hide the fact that the first commit had a
nasty looking user name & email. It didn't even always work..

Now that the initial commits are made using a user specified in the
config file, and which has a suitable default, there's no need to
hide this from users.

Change-Id: I21712e5644733048a65c151c132a172ca05debfb