Create a general process running helper & use it.
I've been meaning to move the contents of the gitExecute functions
to a more general helper for a while now, and I'm planning to use
the new function to replace a number of the shell_exec calls that
are causing stderr dumping to the webserver log (#1445).

No expected functional changes.

Change-Id: I3d10029b7f45fe1b976c963be098b209cef897fc