Don't require a safe marker on 1st arg of gitExecute(Internal)?
This first argument is a path, which if escaped using escapeshellarg
or similar doesn't work properly. It's therefore rather annoying to
tell developers that they need to mark this path safe, given that
if they actually do so the code will break. I feel that this works
better than using the safe marker I introduced, as it will still
pick up on issues in the command.

Change-Id: Idbf8529bc7963fe6ea297f45cc59254d431daea0