Fix #1461: Make layout of the ACE box dynamic.
This improves upon the previous static absolute sizing by adding a
listener for the window resize that adjusts the position of the top
of the ace box, taking into account the location of the previous
element. This provides a solution that is (almost) guarranteed to
show the right thing. Also call said function whenever the page is
shown so that we're sure to be correct, even if the page resize
occured while we weren't shown.

Change-Id: I5165ccf2f3262e98bd23bb21634562c602af8ee2