Fix a bug where team-status showed rejected for empty items.
In the case where the user hasn't submitted a value, the data-store
won't have a value for that field at all, resulting in the isset
check failing. This should have been returning null (as described
in the comment), but wasn't.

This provides a better fix to the change in bccdd8e14f8205.
Also add tests for this case, so that the issue doesn't resurface.

Change-Id: Id744e97b7635e6fc2c92eae28223546b99b29e3c