Separate the UI from its need to have a team id.
There is now quite a lot of functionality in the IDE that doesn't
actually need a team id to work, and it's quite possible that
mentors won't be in a specific team. This commit rearranges how
the UI is launched so that it is less dependant on the user being
in a team. Add a useful message for users that aren't in teams,
and add checks where needed that the user has selected a team,
with an instructional error if they haven't.

This commit also includes quite a bit of restructuring to the
ProjPage object which simplifies the UI setup by moving stuff that
should be internal to the ProjPage into it. Some of these things
are then re-exposed if needed, but the majority aren't.

Change-Id: Ieb3662f449c8f546d1e14fade120b916cdc8f99f