Fix undeletion of multiple files.
This is the same issue (and fix) as was the case in #1570, namely
that we were calling for the repo to be cleaned & updated before
each file, and thus negating most of what we were doing. Also
extend the test I created earlier to check the undelete cycle.

Annoyingly, I've a feeling this is the first test that checks this
sequence of calls.

Change-Id: I7af252243e8edef1643c3abc62d4382e162a530e