Don't always open projects in the proj module.
It turns out that most of the endpoints don't actually need
to use a working copy for their actions, and so don't need an
updated version of it available. Remove the openning of the
project from the module init, and add it into the endpoints
that actually need it.

In the process, move from a module level reference to local
references to the object, which makes it clearer when it's
getting created.

I'm hoping that this will reduce the number of errors that are
caused by lock failures, and initial testing suggests that it does.

Change-Id: Ia43387db721f7678f500194450d75e7660c812ca