Add visual error markers via ace.
This adds error markers in Ace's line-number gutter via its
annotations interface. These are per-line for the moment as
this is both a limitation on the annotations, and of the data
we have available.

These are cleared when we set the value of the ace contents,
meaning that while the clear_errors function is present, it's
not actually needed. Since these are per 'session', each file
gets its own listing of errors (yay).

Change-Id: Ie2edd2f8a8115c9e9e29f8615dcfa64b149df39a