Detect and fail test that fial to load modules.
For some reason the jasmine runner thinks this is a valid thing to
do, and just returns a PASS without running any tests if it happens.
Since this isn't very helpful, especialy since we can't see the output
due to our own wrapping stuff swallowing it for passing tests, ensure
that such failures are reported properly.

Change-Id: Icb02e41e6b672ef9413bb8b9010651edff1544a5