Avoid backend error when re-checking a file.
Previously, the final parameter passed to the check method when
this signal fired was the event being signalled. This didn't appear
to cause any issues, but did result in the backend outputting
an error about passing an object to escapeshellargs.

Adding the additional null to this bind ensures that we're always
passing valid arguments through.

Change-Id: I848117af00ca9113190be65f20c996a1c02f2624