Hide previous project pane when switching teams.
Previously, the information for the previous project, from
the other team, was shown until replaced by the new info.
This change ensure that it's always hidden immediately,
as it is no longer valid.

The particular case where this was an issue was when the new
team had a project of the same name as the one being shown.