Ensure correct statics are output if the combined cache fails.
Previously, if the combination static failed to be written (either
updated or created) for some reason, this failure was silent. In
the case of a failed update, this was very hard to diagnose since
much of the frontend appeared to work.

We now fall back to outputting the direct links if the cache fails
for any reason.