Add read-only checks to the backend.
This prevents any callers from using the backend directly to make
any changes to all parts of a team that they shouldn't have access
to. This includes their projects, files & team status.

Add a helper method to the AuthBackend which provides an exception
in the case that the team isn't writable.

Change-Id: I00c57625070dc94d152889921020890436fe0d58