Fix #1732: Mark errors on a document after navigating to it.
This isn't really the best fix -- ideally, we probably need an error
collection that's separate from the handling of the visual display
of the error, and that should then be subscribed to by both the edit
page and the errors view.

However, since that's a much larger & complicated change, instead
just have the errors page mark the errors on the document it
navigates to. Also have the edit tab postpone actually setting the
error markers until the editor is ready to ensure they appear correctly.

Change-Id: I57f08da4ca57be612535ea0562392a97ca646080