Make team-status images live direclty after review.
This changes the post-review procedure for images such that (if valid)
they are copied into a configurable target folder so that they can
are made live immediately. This also moves the creation of the image
thumbnail to being part of this step, ensuring that the thumbnail
and the reviewed image that go live are the same.

Change-Id: I2d7d3ecc3ab652cc7a8e5aa93a13c99e95870999