Simplify options handling.
Default to an empty object, removing a chunk of null checks, and
cache the actual callback value rather than a boolean telling us
about it (since functions can be used happily in 'if's).
Don't bother to check the type of object passed, as nothing else
does, and it's not a very JavaScript thing to do.

Change-Id: I36cfb0b41c3cc6f5214809f3a112687974548ab0