Avoid updating the ace session content if possible.
When we set the content, ace moves the caret back to the start of
the document, which moves the user out of their context & interrupts
their flow. Both of these are bad and can be avoided in the majority
case where the content that was saved is the same as the current
content.

Change-Id: I6b6613364d6c3c4452116d1825403e7aabccdaf5