Add the start of a module that handles multiple commands.
This will allow the front-end to request multiple things over a
single http request, which ought to be more efficient & allow for
separation of some backend commands which currently violate separations,
but exist to achieve a similar effect in a specialised manner.

This is somewhat experimental at the moment. While it is functional
(and tested for the simple case of independent passing sub-commands)
it is missing any decent error handling, and the sequential command
is untested.

Change-Id: Iad66e7f0704493fbdcbb6d6288d73019fa530819