Reject re-entrant calls.
These aren't supported, as they would need separate layers of caching
the input and output collections, which we don't do at the moment.
While this is something that may be worth adding later, it's not
needed for the moment & thus not supported. Reject them outright
to avoid corruptions of the output data.

Change-Id: Ia17bc97a0bc0d5471c88e62574e4f96809707cf5