Replace assumption of lowercase uids with actual normalisation.
This allows uppercase names to be added in the future without
requiring further IDE changes, and simultaneously ensures that the
IDE can be successfully used by a user that doesn't know the casing
of their username. We now have a normalisation step performed after
a successful login which ensures that the IDE uses the version of
the username that is cannonical, rather than what the user entered.
This ensures that their groups are correctly found, while still
allowing them to login using the wrong case.

Change-Id: I1274cf0d8e5a0958f6f6cff8f86c008374f909fd