Remove duplicate file/log call.
Don't bother calling file/log if the file isn't loaded yet. While
the history that is returned would be valid, it will be fetched again
once the content has loaded. Delaying the request (and removing the
duplicate) reduce the chances that it will clash with the file/get
call for access to the repo.

Change-Id: I35c9fcb4eedff2d3146c04ca0c1e7bed2db4b225