Block project deletion by non-admins.
This isn't supported properly, and would likely cause issues if
teams stared actually using it on the live IDE (the endpoint has
been found and highlighted on the forums). Rather than need to
pick up the pieces, blocking its use should prevent too many issues.

I've left it in place so that the test teardowns which use this can
continue to do so, and to simplifiy the process of implementing this
if/when we do want to support it properly (once we figure out how).

Change-Id: Ide6d69bd4a1c9ce39e39df71885a6bb3ecb2d8e3