Make proc_exec much more useful
When told to catch the result of the process, have it return a dict
style array with the raw information from the process. This allows
it to be used as a wrapper around proc_open without any loss of
information returned.

Previously there was no way to access the stderr of the process,
which lead some code to use proc_open and thus lose out on the
logging provided in proc_exec.

Change-Id: I62095d406e1fdefe4c7eaa9f745c0f06803b44c9