Switch linting to use proc_exec
This is better than having them do their own things with proc_open
as it means we get consistent logging of the processes being launched
and any failures they encounter. This was made possible by the plumbing
in b6b56d78188be67, since we need access to the actual exit code and
stderr stream in one case.

Change-Id: I5ebaf30fe426e6959a5bbaf95135dec99d35758f